// client/volt-client/src/app/layout.tsx
"use client";
import type { Metadata } from "next";
import { Inter } from "next/font/google";
import { createContext, useState } from "react";
import "./globals.css";

const inter = Inter({ subsets: ["latin"] });

const metadata: Metadata = {
  title: "Create Next App",
  description: "Generated by create next app",
};

export const LinesSettingsContext = createContext({
  linesSettings: { orange: true, blue: true, green: true, yellow: true },
  setLinesSettings: (linesSettings: any) => {},
});

export default function RootLayout({
  children,
}: Readonly<{
  children: React.ReactNode;
}>) {
  const [linesSettings, setLinesSettings] = useState({
    orange: true,
    blue: true,
    green: true,
    yellow: true,
  });

  return (
    <html lang="en">
      <LinesSettingsContext.Provider
        value={{ linesSettings, setLinesSettings }}
      >
        <body className={inter.className}>{children}</body>
      </LinesSettingsContext.Provider>
    </html>
  );
}